chore: update response cart

+++ cmd/api/request/AddToCartReq.go
@@ -1,7 +1,7 @@
 package request
 
 type AddToCartRequest struct {
-	ProductID int `json:"product_id" validate:"required"`
-	Quantity  int `json:"quantity" validate:"required"`
-	UserID    int `json:"user_id" validate:"required"`
+	ProductID int64 `json:"product_id" validate:"required"`
+	Quantity  int   `json:"quantity" validate:"required"`
+	UserID    int64 `json:"user_id" validate:"required"`
 }
